Changed to prevent compiler warning.

+++ dev/apps/swrx/EMLTracker.cpp
@@ -187,9 +187,9 @@ void EMLTracker::updateLoop()
    CodeIndex indx = localReplica.codeGenPtr->getIndex() % sync;
    unsigned chips = sync - indx;
    iadCountMax = static_cast<unsigned long>(chips / localReplica.chipsPerTick);
-   if (iadCountMax < static_cast<double>((1/localReplica.tickSize)/2.0)*1e-3 ) 
+   if (iadCountMax < static_cast<unsigned long>((1/localReplica.tickSize)/2.0)*1e-3 ) 
    {
-      iadCountMax += static_cast<double>(1 / localReplica.tickSize )* 1e-3;
+      iadCountMax += static_cast<unsigned long>(1 / localReplica.tickSize )* 1e-3;
    }
 }
